microservices/google-merchant/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-api-python-client